Fixing tests

+++ src/metaphor/api.py
@@ -17,6 +17,7 @@ from metaphor.lrparse.lrparse import CollectionResourceRef
 from metaphor.lrparse.lrparse import RootResourceRef
 from metaphor.lrparse.lrparse import LinkCollectionResourceRef
 from metaphor.lrparse.lrparse import OrderedCollectionResourceRef
+from metaphor.lrparse.lrparse import FieldRef
 from metaphor.schema import CalcField
 from metaphor.schema_factory import SchemaFactory
 from metaphor.updater import Updater
@@ -368,17 +369,18 @@ class Api(object):
                 parent_resource = next(cursor)
 
                 # check for file deletion
-                resource_tree = self._parse_canonical_url(path)
-                resource_field_name = path.rsplit('/', 1)[-1]
-                field_spec = spec.fields[resource_field_name]
+                if type(tree) == FieldRef:
+                    resource_tree = self._parse_canonical_url(path)
+                    resource_field_name = path.rsplit('/', 1)[-1]
+                    field_spec = spec.fields[resource_field_name]
 
-                if field_spec.field_type == 'file':
-                    self.updater.delete_file(
-                        spec.name,
-                        parent_resource['_id'],
-                        resource_field_name,
-                        parent_resource[resource_field_name])
-                    return {}
+                    if field_spec.field_type == 'file':
+                        self.updater.delete_file(
+                            spec.name,
+                            parent_resource['_id'],
+                            resource_field_name,
+                            parent_resource[resource_field_name])
+                        return {}
 
                 parent_spec_name = parent_field_tree.parent_spec.name if parent_field_tree.parent_spec else None
                 return self.updater.delete_resource(spec.name, resource_id, parent_spec_name, field_name)
